+++ Stripe/StripeiOSTests/STPPaymentMethodFunctionalTest.swift
@@ -276,6 +276,21 @@ class STPPaymentMethodFunctionalTest: XCTestCase {
         waitForExpectations(timeout: 5, handler: nil)
     }
 
+    func testCreateMultibancoPaymentMethod() {
+        let client = STPAPIClient(publishableKey: STPTestingDefaultPublishableKey)
+        let billingDetails = STPPaymentMethodBillingDetails()
+        billingDetails.email = "tester@example.com"
+        let params = STPPaymentMethodParams(alma: STPPaymentMethodAlmaParams(), billingDetails: billingDetails, metadata: nil)
+        let expectation = self.expectation(description: "Payment Method create")
+        client.createPaymentMethod(with: params) { paymentMethod, error in
+            XCTAssertNil(error)
+            XCTAssertNotNil(paymentMethod)
+            XCTAssertEqual(paymentMethod?.type, .alma)
+            expectation.fulfill()
+        }
+        waitForExpectations(timeout: 5, handler: nil)
+    }
+
     func fetchPaymentMethods(client: STPAPIClient,
                              customerAndEphemeralKey: STPTestingAPIClient.CreateEphemeralKeyResponse) async throws -> [STPPaymentMethod] {
         try await withCheckedThrowingContinuation { continuation in